+++ NEWS.md
@@ -1,6 +1,7 @@
 
 # vcfR 1.1.0.9000
 
+* `extract.info()` scores missing elements as NA.
 
 
 # vcfR 1.1.0
@@ -26,6 +27,7 @@ This release includes the incorporation of suggestions made by reviewers of the 
 * `create.chromR()` now handles instances with no seq and the annotation position exceeds the greatest VCF POS.
 * `read.vcfR()` now handles tilde expansion.
 * `addID()` populates the non-missing values in the ID column of VCF data by concatenating the chromosome and position. 
+
 
 # vcfR 1.0.0
 Released on CRAN 2016-02-22.